Dev #426 change wording
